Add support for other characters and longer codes (#28)
* Add support for *, +, #, -, . and space in codes
* Repeat passcode for codes with more digits to avoid obvious patterns
* Add support for longer codes using long values and repeating keys

+++ app/src/main/java/se/oscarb/pinapple/Crypto.java
@@ -7,8 +7,10 @@ package se.oscarb.pinapple;
 public interface Crypto {
 
     // Encrypt and decrypt with key and message as int
-    int encrypt(int messagePlainText, int key);
-    int decrypt(int messageEncrypted, int key);
+
+    // Encrypt and decrypt with message as long
+    long encrypt(long messagePlainText, int key);
+    long decrypt(long messageEncrypted, int key);
 
 
 }
